verify if classroom user

+++ backend/src/modules/classrooms/services/DeleteClassroomService.ts
@@ -7,6 +7,7 @@ import IClassroomsRepository from '@modules/classrooms/repositories/IClassroomsR
 
 interface IRequest {
   classroom_id: string;
+  user_id: string;
 }
 
 @injectable()
@@ -16,11 +17,15 @@ class DeleteClassroomService {
     private classroomsRepository: IClassroomsRepository,
   ) {}
 
-  public async execute({ classroom_id }: IRequest): Promise<void> {
+  public async execute({ classroom_id, user_id }: IRequest): Promise<void> {
     const classroom = await this.classroomsRepository.findById(classroom_id);
 
     if (!classroom) {
       throw new AppError('classroom does not exists', 400);
+    }
+
+    if (classroom.user_id !== user_id) {
+      throw new AppError("This classroom doesn't belong to you", 400);
     }
 
     await this.classroomsRepository.deleteById(classroom_id);
